README.md
A WEEX referral code 3AND is a unique code that new users can apply during registration to access special promotions and discounts on the platform. By using the code 3AND, you instantly qualify for a range of incentives, including a generous bonus up to 30,000 USDT. This promotion is designed to give new traders and investors a head start, allowing them to explore the platform with additional funds and enjoy a more rewarding trading experience.